requirements.txt
asgiref==3.7.2
Django==5.0
django-stubs==4.2.7
django-stubs-ext==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
mypy==1.8.0
mypy-extensions==1.0.0
packaging==23.2
psycopg==3.1.16
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
types-pytz==2023.3.1.1
types-PyYAML==6.0.12.12
typing_extensions==4.9.0
uritemplate==4.1.1
